requirements.txt
nextcord==2.5.0
Pillow==9.5.0
python-dotenv==1.0.0
numpy==1.24.3